update: 본인이 작성한 레시피만 수정/삭제 가능하도록

+++ src/main/java/com/sparta/backend/controller/RecipesController.java
@@ -5,6 +5,7 @@ import com.sparta.backend.dto.request.recipes.PostRecipeRequestDto;
 import com.sparta.backend.dto.response.CustomResponseDto;
 import com.sparta.backend.dto.response.recipes.RecipeDetailResponsetDto;
 import com.sparta.backend.dto.response.recipes.RecipeListResponseDto;
+import com.sparta.backend.exception.CustomErrorException;
 import com.sparta.backend.security.UserDetailsImpl;
 import com.sparta.backend.service.RecipesService;
 import com.sparta.backend.service.TagService;
@@ -15,6 +16,7 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
+import java.util.Optional;
 
 @Slf4j
 @RestController
@@ -27,7 +29,7 @@ public class RecipesController {
     @PostMapping("/recipes")
     public CustomResponseDto<?> postRecipe(PostRecipeRequestDto requestDto, @AuthenticationPrincipal UserDetailsImpl userDetails) throws IOException {
         //todo:IOException처리
-        //todo: checkLogin
+        checkLogin(userDetails);
         //레시피 먼저 생성, 등록
         Recipe savedRecipe = recipeService.saveRecipe(requestDto, userDetails);
         //태그 등록할때 저장한 레시피객체도 넣어줌
@@ -40,7 +42,8 @@ public class RecipesController {
     @PutMapping("/recipes/{recipeId}")
     public CustomResponseDto<?> updateRecipe(@PathVariable Long recipeId, PostRecipeRequestDto requestDto, @AuthenticationPrincipal UserDetailsImpl userDetails) throws IOException {
         //todo:IOException처리
-        //todo: checkLogin
+        checkLogin(userDetails);
+        checkOwnership(recipeId, userDetails);
 
         //레시피 업데이트
         Recipe updatedRecipe = recipeService.updateRecipe(recipeId,requestDto,userDetails);
@@ -54,8 +57,9 @@ public class RecipesController {
     //레시피 삭제
     @DeleteMapping("recipes/{recipeId}")
     public CustomResponseDto<?> deleteRecipe(@PathVariable Long recipeId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        //todo: checkLogin
-        //todo: 글쓴사람이 로그인한사람인지 체크
+        checkLogin(userDetails);
+        checkOwnership(recipeId, userDetails);
+
         recipeService.deleteRecipe(recipeId);
         return new CustomResponseDto<>(1, "레시피 삭제 성공", "");
     }
@@ -63,7 +67,7 @@ public class RecipesController {
     //레시피 상세조회
     @GetMapping("recipes/{recipeId}")
     public CustomResponseDto<?> getRecipeDetail(@PathVariable Long recipeId, @AuthenticationPrincipal UserDetailsImpl userDetails){
-        //todo: checkLogin
+        checkLogin(userDetails);
         RecipeDetailResponsetDto recipeDetailResponsetDto = recipeService.getRecipeDetail(recipeId, userDetails);
         return new CustomResponseDto<>(1, "레시피 조회 성공", recipeDetailResponsetDto);
     }
@@ -75,7 +79,7 @@ public class RecipesController {
                                            @RequestParam("isAsc") boolean isAsc,
                                            @RequestParam("sortBy") String sortBy,
                                            @AuthenticationPrincipal UserDetailsImpl userDetails){
-        //todo: checkLogin
+        checkLogin(userDetails);
         page = page-1;
         Page<RecipeListResponseDto> recipesByPage = recipeService.getRecipesByPage(page, size, isAsc, sortBy, userDetails);
         return new CustomResponseDto<>(1, "레시피 리스트 성공", recipesByPage);
@@ -84,8 +88,20 @@ public class RecipesController {
     //레시피 좋아요 등록/취소
     @GetMapping("recipes/likes/{postId}")
     public CustomResponseDto<?> likeRecipe(@PathVariable Long postId, @AuthenticationPrincipal UserDetailsImpl userDetails){
-        //todo: checkLogin
+        checkLogin(userDetails);
         String resultMessage = recipeService.likeRecipe(postId, userDetails.getUser());
         return new CustomResponseDto<>(1, resultMessage,"");
+    }
+
+    private void checkLogin(@AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if (userDetails == null) {
+            throw new CustomErrorException("로그인된 유저만 사용가능한 기능입니다.");
+        }
+    }
+
+    private void checkOwnership(Long recipeId, UserDetailsImpl userDetails){
+        Optional<Recipe> recipe = recipeService.findById(recipeId);
+        if(!recipe.get().getUser().getEmail().equals(userDetails.getUser().getEmail())) throw new CustomErrorException("본인의 게시물만 수정,삭제 가능합니다.");
+
     }
 }
